chapter-20-capstone-streamlit/requirements.txt
-e streamlit>=1.30
openai>=1.0
python-dotenv>=1.0
